chalice_deployment/requirements.txt
pycryptodome==3.3.1
python-twitter
requests
warrant